refactor(cli): move telemetry code to shared @fern-api/cli-telemetry package
