+++ packages/cli/cli-telemetry/src/index.ts
@@ -1,0 +1,2 @@
+export { getFernRunId, getOrCreateFernRunId } from "./fernRunId.js";
+export { setSentryFernRunIdTag, setSentryGithubRunIdTag, setSentryRunIdTags } from "./sentryRunId.js";
