+++ packages/cli/cli-telemetry/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "@fern-api/configs/tsconfig/main.json",
+  "compilerOptions": {
+    "outDir": "lib",
+    "rootDir": "src"
+  },
+  "include": ["./src/**/*"]
+}
