+++ packages/cli/cli-v2/src/cli.ts
@@ -1,3 +1,4 @@
+import { getOrCreateFernRunId } from "@fern-api/cli-telemetry";
 import type { Argv } from "yargs";
 import yargs from "yargs";
 import { hideBin } from "yargs/helpers";
@@ -12,7 +13,6 @@ import { addOrgCommand } from "./commands/org/index.js";
 import { addSdkCommand } from "./commands/sdk/index.js";
 import { addTelemetryCommand } from "./commands/telemetry/index.js";
 import { GlobalArgs } from "./context/GlobalArgs.js";
-import { getOrCreateFernRunId } from "./fernRunId.js";
 import { Version } from "./version.js";
 
 export async function runCliV2(argv?: string[]): Promise<void> {
