+++ packages/cli/cli-v2/src/index.ts
@@ -1,3 +1,1 @@
 export { runCliV2 } from "./cli.js";
-export { getFernRunId, getOrCreateFernRunId } from "./fernRunId.js";
-export { setSentryFernRunIdTag, setSentryGithubRunIdTag, setSentryRunIdTags } from "./telemetry/sentryRunId.js";
