+++ packages/cli/cli/package.json
@@ -57,6 +57,7 @@
     "@fern-api/cli-logger": "workspace:*",
     "@fern-api/cli-migrations": "workspace:*",
     "@fern-api/cli-source-resolver": "workspace:*",
+    "@fern-api/cli-telemetry": "workspace:*",
     "@fern-api/cli-v2": "workspace:*",
     "@fern-api/configs": "workspace:*",
     "@fern-api/configuration": "workspace:*",
